refactor(glsl): move the volume flattening out of the translator
A pack that ships a three dimensional texture has every declaration of it
rewritten onto a flat atlas and every read sent through a helper, which is
one pass with a table of what it found and two counts of what it could not
move. That pass, the reading of a name through its macro aliases and the
helper the header writes now live in VolumeFlattening, handed the tokens,
the unit, the volumes the pack declared and the two macro tables. The
translator holds one and asks it for what the header and the notes need.
